Refactor into 'CustomFormValidationAction'
