+++ actions/actions.py
@@ -26,11 +26,11 @@ from actions.parsing import (
 )
 from actions.profile import create_mock_profile
 
+from actions.custom_forms import CustomFormValidationAction
+
 
 logger = logging.getLogger(__name__)
 
-
-MAX_VALIDATION_FAILURES = 2
 
 NEXT_FORM_NAME = {
     "pay_cc": "cc_payment_form",
@@ -44,169 +44,6 @@ FORM_DESCRIPTION = {
     "transfer_money_form": "money transfer",
     "transaction_search_form": "transaction search",
 }
-
-
-async def shared_validate_continue_form(value: Text) -> Dict[Text, Any]:
-    """Validates value of 'continue_form' slot, which is identical for all the forms"""
-    if value == "yes":
-        return {"continue_form": value}
-
-    if value == "no":
-        # This will activate rule 'Submit ---_form' to cancel the operation
-        return {
-            "requested_slot": None,
-            "confirm": "no",
-            "continue_form": value,
-        }
-
-    # The user's answer was not valid. Just re-set it to None.
-    return {"continue_form": None}
-
-
-class MyFormValidationAction(FormValidationAction):
-    """Tracks repeated validation failures"""
-
-    # making this an abstractmethod avoids registering a custom action
-    @abstractmethod
-    def name(self) -> Text:
-        pass
-
-    async def validate(
-        self,
-        dispatcher: CollectingDispatcher,
-        tracker: Tracker,
-        domain: Dict,
-    ) -> List[EventType]:
-        """In addition to regular slot validation,
-        - tracks repeated validation failures
-        - when MAX_VALIDATION_FAILURES reached:
-          - explains the slot, if an explain_{slot} method is provided
-          - asks user to continue or not
-        """
-        events = []
-
-        if not tracker.get_slot("continue_form"):
-            events.append(SlotSet("continue_form", "yes"))
-
-        events.extend(await super().validate(dispatcher, tracker, domain))
-
-        events.extend(
-            await self.repeated_validation_failures(dispatcher, tracker, domain, events)
-        )
-
-        return events
-
-    async def repeated_validation_failures(
-        self,
-        dispatcher: CollectingDispatcher,
-        tracker: Tracker,
-        domain: Dict,
-        events: List[EventType],
-    ) -> List[EventType]:
-        """Updates the slot repeated_validation_failures, and sets required form slot
-        `continue_form` to None when the threshold is reached.
-
-        This will trigger utter_ask_{form}_continue_form, asking the user if they want
-        to continue with this form or not.
-        """
-        rvf_events = []
-        requested_slot = tracker.get_slot("requested_slot")
-
-        # Only do this while form is asking for a certain slot
-        if not requested_slot:
-            return rvf_events
-
-        # Skip if validate_{slot} turned off the form by setting requested_slot to None
-        for event in events:
-            if (
-                event["event"] == "slot"
-                and event["name"] == "requested_slot"
-                and not event["value"]
-            ):
-                rvf_events.append(SlotSet("repeated_validation_failures", 0))
-                return rvf_events
-
-        rvf = tracker.get_slot("repeated_validation_failures")
-        if rvf:
-            rvf = int(rvf)
-        else:
-            # initialize counter to 0
-            rvf = 0
-
-        # check if validation of the requested_slot failed
-        validation_failed = False
-        for event in events:
-            if (
-                event["event"] == "slot"
-                and event["name"] == requested_slot
-                and not event["value"]
-            ):
-                validation_failed = True
-                break
-
-        # keep track of repeated validation failures
-        if validation_failed:
-            rvf += 1
-        else:
-            rvf = 0
-
-        if rvf >= MAX_VALIDATION_FAILURES:
-            rvf_events.extend(
-                await self.explain_requested_slot(dispatcher, tracker, events)
-            )
-            # reset counter
-            rvf = 0
-
-            # Triggers 'utter_ask_{form}_continue_form'
-            rvf_events.append(SlotSet("continue_form", None))
-
-        rvf_events.append(SlotSet("repeated_validation_failures", rvf))
-        return rvf_events
-
-    async def explain_requested_slot(
-        self,
-        dispatcher: "CollectingDispatcher",
-        tracker: "Tracker",
-        domain: "DomainDict",
-    ) -> List[EventType]:
-        """Explains requested slot by calling an explain function for the slot.
-
-        Args:
-            dispatcher: the dispatcher which is used to
-                send messages back to the user.
-            tracker: the conversation tracker for the current user.
-            domain: the bot's domain.
-        Returns:
-            `SlotSet` events for the explained slot (Optional).
-        """
-        slot_name = tracker.get_slot("requested_slot")
-        if not slot_name:
-            return []
-
-        slot_value = tracker.get_slot(slot_name)
-
-        function_name = f"explain_{slot_name.replace('-','_')}"
-        explain_func = getattr(self, function_name, None)
-
-        if not explain_func:
-            logger.debug(
-                f"Skipping explanation for `{slot_name}`: there is no explanation "
-                "function specified."
-            )
-            return []
-
-        slots = {}
-        if utils.is_coroutine_action(explain_func):
-            explanation_output = await explain_func(
-                slot_value, dispatcher, tracker, domain
-            )
-        else:
-            explanation_output = explain_func(slot_value, dispatcher, tracker, domain)
-
-        if explanation_output:
-            slots.update(explanation_output)
-
-        return [SlotSet(slot, value) for slot, value in slots.items()]
 
 
 class ActionPayCC(Action):
@@ -260,22 +97,12 @@ class ActionPayCC(Action):
         return [SlotSet(slot, value) for slot, value in slots.items()]
 
 
-class ValidatePayCCForm(MyFormValidationAction):
+class ValidatePayCCForm(CustomFormValidationAction):
     """Validates Slots of the cc_payment_form"""
 
     def name(self) -> Text:
         """Unique identifier of the action"""
         return "validate_cc_payment_form"
-
-    async def validate_continue_form(
-        self,
-        value: Text,
-        dispatcher: CollectingDispatcher,
-        tracker: Tracker,
-        domain: Dict[Text, Any],
-    ) -> Dict[Text, Any]:
-        """Validates value of 'continue_form' slot"""
-        return await shared_validate_continue_form(value)
 
     async def validate_amount_of_money(
         self,
@@ -460,7 +287,7 @@ class ActionTransactionSearch(Action):
         return [SlotSet(slot, value) for slot, value in slots.items()]
 
 
-class ValidateTransactionSearchForm(MyFormValidationAction):
+class ValidateTransactionSearchForm(CustomFormValidationAction):
     """Validates Slots of the transaction_search_form"""
 
     def name(self) -> Text:
@@ -481,16 +308,6 @@ class ValidateTransactionSearchForm(MyFormValidationAction):
                 events.append(SlotSet("requested_slot", "vendor_name"))
 
         return events
-
-    async def validate_continue_form(
-        self,
-        value: Text,
-        dispatcher: CollectingDispatcher,
-        tracker: Tracker,
-        domain: Dict[Text, Any],
-    ) -> Dict[Text, Any]:
-        """Validates value of 'continue_form' slot"""
-        return await shared_validate_continue_form(value)
 
     async def validate_search_type(
         self,
@@ -570,22 +387,12 @@ class ActionTransferMoney(Action):
         return [SlotSet(slot, value) for slot, value in slots.items()]
 
 
-class ValidateTransferMoneyForm(MyFormValidationAction):
+class ValidateTransferMoneyForm(CustomFormValidationAction):
     """Validates Slots of the transfer_money_form"""
 
     def name(self) -> Text:
         """Unique identifier of the action"""
         return "validate_transfer_money_form"
-
-    async def validate_continue_form(
-        self,
-        value: Text,
-        dispatcher: CollectingDispatcher,
-        tracker: Tracker,
-        domain: Dict[Text, Any],
-    ) -> Dict[Text, Any]:
-        """Validates value of 'continue_form' slot"""
-        return await shared_validate_continue_form(value)
 
     async def validate_PERSON(
         self,
